refactor: split out csv generation in gen_csv.rs
