+++ backend/src/lib.rs
@@ -1,29 +1,19 @@
 pub mod db;
+mod gen_csv;
 mod rest;
 mod schema;
 mod stats;
 
-use crate::db::TableInfo;
 use clap::Parser;
 use diesel::SqliteConnection;
 use log::{debug, error, info, warn};
 use rayon::iter::IntoParallelRefIterator;
 use rayon::iter::ParallelIterator;
 use stats::Stats;
-use std::io::Write;
 use std::sync::mpsc;
 use std::sync::{Arc, Mutex};
 use std::{error, fmt, io, thread};
 
-const METRIC_TABLES: [&str; 5] = [
-    "block_stats",
-    "tx_stats",
-    "script_stats",
-    "input_stats",
-    "output_stats",
-];
-const COLUMN_NAMES_THAT_ARENT_METRICS: [&str; 6] =
-    ["height", "date", "version", "nonce", "bits", "pool_id"];
 const DATABASE_BATCH_SIZE: usize = 100;
 
 // Don't fetch (and process) the most recent blocks to be safe
@@ -322,47 +312,7 @@ pub fn write_csv_files(
     csv_path: &str,
     connection: Arc<Mutex<SqliteConnection>>,
 ) -> Result<(), MainError> {
-    let connection = Arc::clone(&connection);
-    let mut conn = connection.lock().unwrap();
-    info!("Generating date.csv file...");
-    let date_column = db::date_column(&mut conn);
-    let mut date_file = std::fs::File::create(format!("{}/date.csv", csv_path))?;
-    let date_content: String = date_column
-        .iter()
-        .map(|row| format!("{}\n", row.date))
-        .collect();
-    date_file.write_all("date\n".as_bytes())?;
-    date_file.write_all(date_content.as_bytes())?;
-
-    for table in METRIC_TABLES.iter() {
-        let columns = db::list_column_names(&mut conn, table)?;
-
-        // filter out columns that aren't metrics and we don't want to create csv files for
-        let columns_filtered: Vec<&TableInfo> = columns
-            .iter()
-            .filter(|col| !COLUMN_NAMES_THAT_ARENT_METRICS.contains(&&col.name[..]))
-            .collect();
-
-        for column in columns_filtered.iter().map(|col| col.name.clone()) {
-            info!("Generating metrics for '{}' in table '{}'.", column, table);
-            let avg_and_sum = db::column_sum_and_avg_by_date(&mut conn, &column, table);
-
-            let mut avg_file = std::fs::File::create(format!("{}/{}_avg.csv", csv_path, column))?;
-            let avg_content: String = avg_and_sum
-                .iter()
-                .map(|aas| format!("{:.4}\n", aas.avg))
-                .collect();
-            avg_file.write_all(format!("{}_avg\n", column).as_bytes())?;
-            avg_file.write_all(avg_content.as_bytes())?;
-
-            let mut sum_file = std::fs::File::create(format!("{}/{}_sum.csv", csv_path, column))?;
-            let sum_content: String = avg_and_sum
-                .iter()
-                .map(|aas| format!("{}\n", aas.sum))
-                .collect();
-            sum_file.write_all(format!("{}_sum\n", column).as_bytes())?;
-            sum_file.write_all(sum_content.as_bytes())?;
-        }
-    }
+    gen_csv::date_csv(csv_path, connection.clone())?;
+    gen_csv::metrics_csv(csv_path, connection.clone())?;
     Ok(())
 }
